pyproject.toml
[build-system]
requires = ["poetry>=1.8.2"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ska-dlm-client"
version = "1.1.0"
description = "Client(s) for the SKA Data Lifecycle Manager"
authors = ["Mark Boulton <mark.boulton@uwa.edu.au>"]
repository = "https://gitlab.com/ska-telescope/ska-dlm-client"
documentation = "https://developer.skao.int/projects/ska-dlm-client/en/latest/"
license= "BSD-3-Clause"
readme = "README.md"
packages = [{include="ska_dlm_client", from="src"}]

[tool.poetry.scripts]
dlm-directory-watcher = "ska_dlm_client.directory_watcher.main:main"
dlm-kafka-watcher = "ska_dlm_client.kafka_watcher.main:main"
dlm-register-storage-location = "ska_dlm_client.register_storage_location.main:main"
dlm-startup-verification = "ska_dlm_client.startup_verification.main:main"

[[tool.poetry.source]]
name = 'skao'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'
priority = "primary"

[[tool.poetry.source]]
name = "PyPI"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.10"
aiokafka = "0.10.0"
pydantic = "^2.9.2"
python-dateutil = "^2.9.0.post0"
python-benedict = "^0.33.2"
watchfiles = "^0.24.0"
pyyaml = "^6.0.2"
ska-sdp-dataproduct-metadata = "^0.7.0"
watchdog = "^6.0.0"
crc32c = "^2.7.1"
typer = "^0.19.2"

[tool.poetry.group.test.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.12.0"
pylint = "^3.2.7"
pylint-junit = "^0.3.2"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
flake8-docstrings = "^1.7.0"
flake8-rst-docstrings = "^0.3.0"
flake8-pyproject = "^1.2.3"
requests-mock = "^1.12.1"
pytest-mock = "^3.14.0"

[tool.poetry.group.docs.dependencies]
sphinx = ">=7.2.0"
ska-ser-sphinx-theme = ">=0.1.1"
sphinx-autodoc-typehints = ">=1.23.0"
pyyaml = "^6.0.2"
docconvert = "^2.2.0"
myst-parser = "^4.0.0"

[tool.pytest.ini_options]
addopts = "-v"

[tool.flake8]
docstring-style = "google"
rst-directives = "literalinclude"
ignore = ['F401', 'W503']
max-complexity = 10
max-line-length = 180

# exclude for auto generated code
exclude = ["src/ska_dlm_client/openapi", "tests/openapi"]

[tool.pylint.messages_control]
disable = [
    "duplicate-code",
    "fixme",
]
enable = ["useless-suppression"]

[tool.pylint.MASTER]
ignore = "openapi"

[tool.isort]
line_length = 99

[tool.black]
line-length = 99

[tool.pylint.format]
max-line-length = 99

[tool.pylint.similarities]
min-similarity-lines = 7
ignore-imports = "yes"
